requirements.txt
asgiref>=3.2.7
cffi==1.14.0
cryptography>=3.3.2
Django>=3.1.6
django-geojson==3.0.0
django-leaflet==0.26.0
django-rest-knox==4.1.0
djangorestframework>=3.11.2
django-autocomplete-light>=3.8.2
django-select2
django-ckeditor
django-filter
django-bootstrap3
django-bootstrap-form
django-crequest
jsonfield==3.1.0
mysqlclient==1.4.6
Pillow>=8.1.1
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
